Created Initial Files
Creates some of the files needed in order to create the game. Each has a comment stating what they are for.

Board: Map
Card: The Cards in the Hand
Game: General game logic and flow
Main: Used to initiate the game with the correct settings
Player: Player commands and actions
Room: Each room on the Board
Tile: Each moveable tile