Show Human Player in Room
DISCLAIMER
I dont know how the AI player uses the door but it doesn't call the activate function so doesn't work yet.
It should activate all tiles it goes on to instead of just showing it there, otherwise the player can go on a tile when it shouldnt be able to.

The tile image can now be set in place in addOccupier for the tile. This is how all players should be added to a tile. Remove occupier removes the image, as opposed to the activate.

---

Board:
~ rooms --> ArrayList, dont need exact position
~ moved creating rooms and doors, wasnt entirely needed but oh well
~ add to room now array
+ add playerIndicator to add coordinates to rooms
+ {room}Players: Room icon coordinates
+ getRooms

Door:
- removed unused import
~ remove adding door occupier, instead move into room

Place:
- Unnecessary spaces
+ removeOccupier now also removes image
- activate no longer removes image
+ activate removes occupier from old space and adds to new one

Room:
- Handling of player room icons done in GameController
+ get the coordinates of player icons

Tile:
+ Empty Javadoc space for getKetFromValue (idk how used)

NPC:
- removed unused import

Player:
~ getIsInRoom now isInRoom

GameController:
+ board and game now static final and determined upon declaration
+ get HashMap playerMarkers to store all ImageView for room icons. Key: RoomName, Value: ImageView[], where the index is in player order (scarlett = 0, mustard = 1, etc)
+ show/hidePlayerRoom: player icon in spec room
!!! AI NEEDS TO ACTIVATE TILE NOT JUST SHOW ON THEM